Rename requests - "first" and "other" rather than "primary" and "additional".
"Primary" implied there was something special about the first request, it's treated exactly like the others, it's just there so we can guarantee we have at least one request, as required for a valid packet. "Other" is shorter than "additional" and has less of an implication that they are secondary in some way.