Add new connection states (disconnected, errored) and handle state transitions.
Add error method to move directly to errored state (e.g. when user of state can't send a packet, rendering state invalid).

Allow the newly generated packets in events to live longer than the received packet, by adding a lifetime to event enum. This in turn means that users of the client state can receive a packet with lifetime 'a, pass it to client state and get it back again as a publish, then dispose of the publish method, while still keeping the puback message with a longer lifetime. This is specific to publish at the moment, but may be useful for QoS 2 pub messages in future as well.