Change "quality of service" shortening in camel case names from "QoS" to "Qos".
It's not completely clear what the correct approach is, but Rust treats upper case acronyms/initialisms/shortenings as one word, so e.g. "TCP" becomes "Tcp" in `TcpStream`. Therefore we'll treat "QoS" as being essentially one single acronym as if it were "QOS". This is at least consistent, and also does read better when you get used to it.